Always init tutor links
