+++ app/javascript/packs/pythia_submission.js
@@ -1,6 +1,3 @@
-import { initPythiaSubmissionShow } from "pythia_submission.ts";
+import { initFileViewers } from "file_viewer";
 
-window.dodona.initPythiaSubmissionShow = initPythiaSubmissionShow;
-
-// will automatically bind to window.iFrameResize()
-require("iframe-resizer"); // eslint-disable-line no-undef
+window.dodona.initFileViewers = initFileViewers;
